+++ notion_df/core/variable.py
