+++ notion_df/misc.py
@@ -9,7 +9,7 @@ from uuid import UUID
 from typing_extensions import Self
 
 from notion_df.core.serialization import DualSerializable
-from notion_df.object.constant import BlockColor, OptionColor
+from notion_df.constant import BlockColor, OptionColor
 
 if TYPE_CHECKING:
     from notion_df.entity import Block, Database, Page, Workspace
