+++ notion_df/request/search.py
@@ -3,8 +3,8 @@ from typing import Any, Literal, Union
 
 from notion_df.core.data import EntityData
 from notion_df.core.request import RequestSettings, Version, Method, PaginatedRequestBuilder
-from notion_df.object.data import DatabaseData, PageData
-from notion_df.object.sort import TimestampSort
+from notion_df.data import DatabaseData, PageData
+from notion_df.sort import TimestampSort
 from notion_df.core.collection import DictFilter
