+++ notion_df/rich_text.py
@@ -3,13 +3,13 @@ from __future__ import annotations as _
 import functools
 from abc import ABCMeta, abstractmethod
 from dataclasses import dataclass, field
-from typing import Optional, Any, Literal, final, Iterable, cast
+from typing import Optional, Any, Literal, Iterable, cast
 from uuid import UUID
 
 from typing_extensions import Self
 
 from notion_df.core.serialization import DualSerializable, deserialize, serialize
-from notion_df.object.misc import DateRange, Annotations
+from notion_df.misc import DateRange, Annotations
 from notion_df.core.collection import FinalDict
 
 span_registry: FinalDict[tuple[str, ...], type[Span]] = FinalDict()
