+++ notion_df/sort.py
@@ -3,7 +3,7 @@ from dataclasses import dataclass
 from typing import Literal, Any
 
 from notion_df.core.serialization import Serializable
-from notion_df.object.constant import TimestampName
+from notion_df.constant import TimestampName
 
 Direction = Literal['ascending', 'descending']
