+++ test/notion_df/script/database_query.py
@@ -2,7 +2,7 @@ from datetime import date
 from pprint import pprint
 
 from notion_df.entity import Database
-from notion_df.object.misc import DateRange
+from notion_df.misc import DateRange
 from notion_df.property import DateProperty, URLProperty
 
 if __name__ == '__main__':
