+++ test/notion_df/util/test_serialization.py
@@ -1,7 +1,7 @@
 from datetime import datetime, date
 
 from notion_df.core.serialization import deserialize_datetime, serialize_datetime
-from notion_df.variable import my_tz
+from notion_df.core.variable import my_tz
 
 
 def test_serde_datetime():
