+++ workflow/action/migration_backup.py
@@ -7,7 +7,7 @@ from loguru import logger
 
 from notion_df.core.request import RequestError
 from notion_df.entity import Page, Database
-from notion_df.object.data import PageData
+from notion_df.data import PageData
 from notion_df.property import RelationProperty, PageProperties, RelationDatabasePropertyValue, \
     DualRelationDatabasePropertyValue
 from workflow.action.match import get_earliest_date
